+++ tests/parser/test_load_languages.py
@@ -8,8 +8,8 @@ def test():
     assert languages == {
         "a": {"k": "ඣ", "l": "3", "x": "1", "y": "3"},
         "b": {"k": "2", "l": "ผ", "x": "2", "y": "4"},
-        "en_us": {"cat": "Cat"},
-        "de_de": {"spoon": "Löffel"},
+        "en_us": {"cat": "Cat", "spoon": ""},
+        "de_de": {"cat": "", "spoon": "Löffel"},
     }
 
 
@@ -18,6 +18,6 @@ def test_prefix_filename():
     assert languages == {
         "a": {"a.x": "1", "a.y": "3", "unicode.k": "ඣ", "unicode.l": "3"},
         "b": {"a.x": "2", "a.y": "4", "unicode.k": "2", "unicode.l": "ผ"},
-        "en_us": {"missing_translations.cat": "Cat"},
-        "de_de": {"missing_translations.spoon": "Löffel"},
+        "en_us": {"missing_translations.cat": "Cat", "missing_translations.spoon": ""},
+        "de_de": {"missing_translations.cat": "", "missing_translations.spoon": "Löffel"},
     }