+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Renamed `--pretty-print`/`-p` to `--minify`/`-m`
 - Renamed `--prefix-filename`/`-n` to `--prefix-identifiers`/`-p`
 - Prefix now prefixes relative path to source
+- Missing translations are no longer skipped. Instead the default to ""
 
 ## [1.2.0] - 2021-03-17
 ### Added