+++ CHANGELOG.md
@@ -6,12 +6,13 @@
 - Source can now be file or directory
 - Accepts multiple sources
 - Improved logging configurability
+- Added support for csv comments
 
 ### Breaking
 - Changed output dir to `--out`/`-o` option
 - Renamed `--pretty-print`/`-p` to `--minify`/`-m`
 - Renamed `--prefix-filename`/`-n` to `--prefix-identifiers`/`-p`
-- Prefix now prefixes relative path to source
+- Prefix option now prefixes relative path to source
 - Missing translations are no longer skipped. Instead the default to ""
 
 ## [1.2.0] - 2021-03-17